project_cemphris/requirements.txt
django==4.2.13
python-dotenv==1.0.1
dj-database-url==2.1.0
gunicorn==21.2.0
channels==4.1.0
channels-redis==4.2.0
django-cors-headers==4.3.1
djangorestframework
djangorestframework_simplejwt
psycopg2-binary==2.9.9
# django-payments-razorpay
drf-yasg
setuptools==70.0.
django-rest-swagger
firebase-admin
Pillow
celery==5.3.6
django-celery-beat==2.6.0
django-redis
redis
celery[redis]
flower
uvicorn==0.30.1
uvicorn[standard]
# graphviz
# django-extensions